+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Changelog
 - Use vector drawable for adaptive app icon
 - Fix single sms being sent to webhook over multiple requests
 - Log broadcast receiver call to external log file
+- Add option to export app logs file
 
 ## 0.1
 - Initial version